requirements.txt
python-socketio==5.0.3
PyDirectInput
PYXInput
pyautogui
flask-socketio
Flask
websockets
pyvjoy
Flask-Sockets